add deled button & moving functions
